UsbReader/requirements.txt
psycopg2==2.6.1
pyserial==2.7
